Create EventTypeUtil and common events
Create the BasicEvent and CommonEvent event classes as examples
of AbstractEvent child classes.

Create EventTypeUtil utility that can generate event names on the
static properties of a class. This will make sure the event names
comform to a standard and that all event types are namespaced
with the Class name.