Fix removal of event listeners
Make sure that event listeners that are removed are no longer
executed, even for the current event loop.

Make sure that removing event listeners does not break the
current event loop by creating a local copy of the array
with listeners during execution.

Renamed 'type' parameters to 'eventType'. We don't use type
because it is a reserved keyword in TS.