Rename listener to handler
Renamed the Listener type alias to EventHandler.
Renamed the listener arguments of EventDispatcher and AbstractEvent
methods to handler.

We now call the binding of an event to a function that will be called
on dispatch a 'listener', and the function itself a 'handler'.
Previously both were called 'listener'. This distinguishment is not
officially in the definition of the words 'listener' and 'handler',
but it helps us separate the concerns when reading the code.