Create BaseEvent and createEventType util
Create BaseEvent class which is an implementation of AbstractEvent
with a generic data property and the type property narrowed from
string to the generic TType. This allows us to use string literal
union types as event type.

Create a utility createEventType that will generate a child class
of BaseEvent that has strictly type event types and data types.

Remove the old CommonEvent and BasicEvent. As it is now easier to
create event types using the createEventType util, it is preferable
to create your own event type instead of using a generic class.