Strongly type EventDispatcher
Update EventDispatcher to be generic, allowing (optionally) to
restrict the types of events that EventDispatcher is allowed to
dispatch. This will also automatically type the 'event' argument
of event handlers.

Remove temporary test code

Updated prettier so it understands new Typescript syntax

Updated typescript to stable version.

Removed IEventDispatcher and IEvent because it was degrading the
strict types