Explicitly set .data in IsomorphicBaseEvent
Set the .data property again (even though it's already defined
in the base class). This is related to this issue:
https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-7.html#the-usedefineforclassfields-flag-and-the-declare-property-modifier

It seems to be that things are running fine when compiling, but
if running in jest or ts-node, the behavior described in
useDefineForClassFields will be used instead. This will cause
the specialization of the data property to a sub-type in
IsomorphicBaseEvent to actually re-declare the property as
undefined.

Closes #50